Potential fix for code scanning alert no. 12: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to avoid overflow. This can be achieved by casting one of the operands to unsigned long before performing the multiplication. This way, the multiplication will be done using the unsigned long type, which has a larger range and can accommodate larger values without overflowing.

Specifically, we will cast rgroup to unsigned long before performing the multiplication. This change will be made on line 160 of the file libjpeg-turbo-2.0.2/jdmainct.c.

+++ libjpeg-turbo-2.0.2/jdmainct.c
@@ -157,7 +157,7 @@ alloc_funny_pointers(j_decompress_ptr cinfo)
      */
     xbuf = (JSAMPARRAY)
       (*cinfo->mem->alloc_small) ((j_common_ptr)cinfo, JPOOL_IMAGE,
-                                  2 * (rgroup * (M + 4)) * sizeof(JSAMPROW));
+                                  (unsigned long)2 * rgroup * (M + 4) * sizeof(JSAMPROW));
     xbuf += rgroup;             /* want one row group at negative offsets */
     main_ptr->xbuffer[0][ci] = xbuf;
     xbuf += rgroup * (M + 4);
